refactor(inventory): 발주 파일을 procurement/ 로 이동 (#724 항목 5-a)
파일 이동과 모듈 배선만. 로직 변경 0줄, API 경로 변경 0건
(`@Controller('purchase-orders')` 그대로 → admin-web·Tauri 변경 0).
경계는 ADR-0032 결정 4 가 소유하며 procurement.module.ts docstring 이 그걸 인용한다.

계획에 없던 것 하나: 공유 순수 헬퍼 2개가 양쪽 BC 에서 쓰이고 있었다.
- earliest-expected-date.ts — inbound.service.ts:428 이 earliestExpectedDate 를 쓴다
- calendar-date.validator.ts — inbound/dto/simple-inbound.dto.ts 가 쓴다
procurement/ 에 두면 inbound → procurement 역방향 의존이 생겨 ADR-0032 결정 4 를
어기므로 중립 지대(shared/dates/, shared/dto/)로 보냈다.

검증: type-check 0 · jest 489 suite 통과(기준선과 동일) · nest build core 성공
· procurement→inbound 호출 2개(ensurePlanForPurchaseOrder, addInboundPlanItems)
· inbound→procurement 역방향 0곳 · procurement→warehouse-transfer 제품코드 0곳
