+++ apps/core/src/modules/inventory/inventory.module.ts
@@ -3,6 +3,7 @@ import { EventsModule } from '@app/events';
 import { INVENTORY_STREAM } from '@packages/event-contracts';
 import { CoreInventoryModule } from './core/inventory.module';
 import { InboundModule } from './inbound/inbound.module';
+import { ProcurementModule } from './procurement/procurement.module';
 import { MovementModule } from './movement/movement.module';
 import { StocktakingModule } from './stocktaking/stocktaking.module';
 import { SuppliersModule } from './suppliers/suppliers.module';
@@ -30,6 +31,7 @@ import { WarehouseTransferModule } from './warehouse-transfer/warehouse-transfer
     SkuCatalogModule,
     WarehouseTransferModule,
     InboundModule,
+    ProcurementModule,
     MovementModule,
     StocktakingModule,
     SuppliersModule,
